fix: make overall percentage a float on all PHP versions and drop unverifiable view-string annotation
- PHP's `/` operator returns an int when the division is exact, and on
  PHP 8.3 min() then propagates that int, so the overall percentage was
  int 100 instead of float 100.0. An explicit float cast makes the type
  deterministic across PHP versions.
- The `@var view-string` annotation on $view only validates when
  larastan can boot an app that has the package's view namespace
  registered, which is true locally but not in CI. The property is a
  plain string; the view name is covered by the rendering tests.

+++ src/Tables/Columns/MultiProgressColumn.php
@@ -44,9 +44,6 @@ use function Filament\Support\get_component_color_classes;
  */
 class MultiProgressColumn extends Column
 {
-    /**
-     * @var view-string
-     */
     protected string $view = 'filament-advanced-components::tables.columns.multi-progress-column';
 
     /**
@@ -615,7 +612,10 @@ class MultiProgressColumn extends Column
             ];
         }
 
-        $overallPercentage = $isEmpty ? 0.0 : min($sum / $total * 100, 100.0);
+        // The explicit float cast matters: PHP's `/` returns an int when the
+        // division is exact, and min() may then return that int on some PHP
+        // versions.
+        $overallPercentage = $isEmpty ? 0.0 : min((float) ($sum / $total * 100), 100.0);
 
         return [
             'segments' => $processed,
